Introduce contrast in lightness calculations
In particular, make palettes with higher contrast spread out accent
colors more (so that e.g. lightness difference between blue and cyan is
bigger). Generally we want to spread out accent colors, but the limit is
that the darkest colors must still have enough contrast with the
background - so it makes sense to take advantage of that in versions
with higher contrast.

This also makes secondary content color (used for comments) a bit
darker, for better discernibility from foreground. However, it is still
slightly more contrasting with the background than in Solarized.