Version 0.8: add monotone for highlighting
I want to have different colors for "strong" and "weak" highligthing.
The first can be used e.g. for selected text, and the second for marking
the line with the cursor.

Since standard terminal color codes are limited, I'm reassigning "white"
color to the previous "bright black" (used for comments and secondary
content). This *may* cause problems if some programs expect this color
to be the same as foreground - will have to check in practice.