Decrease comment and highlight contrast
Previous comments were not distinct enough from regular foreground. The
idea here is that a comment block shoudld be immediately distinguishable
from a block of code by the sheer difference in lightness, so that even
if you are far away from the screen you will tell what is commented out
by the lightness of text.

I tried setting comment lightness to 3/5 of foreground lightness, but
that was too low (in particular it was impossible to find appropriate
highlight lightness that wouldn't make highlighted comments very hard
to read).

As for background highlight colors, they were both too strong -
highlighting the text negatively affected readability. Since I decreased
their lightness difference compared to foreground, the default
saturation correction had to be decreased as well.