More dynamic accents in selenized light
I overdid the lighness limitations for accent colors previously - the
result was that yellow and cyan were much too dark and therefore ugly.
It's also harder to tell the hue of an accent color if it's too dark.

With these settings, minimum contrast between accent colors and
background in selenized light is 37, just as in solarized light
(compared to 35 in version 0.8.1 and 39.5 in version 0.8.2). I think
it's a good compromise between good-looking accents and enough contrast.