Increase contrast of grey variant
I've noticed that the grey version looks best when it has more contrast
and is overall lighter. It then works pretty well when placed
side-to-side with black-on-white windows like webpages in a browser.

For some reason increasing the lightness and contrast worked well for
grey variant and not for selenized medium.