Default iTerm color space is not AppleRGB
it's Generic RGB actually. And when exporting a color preset, it's saved
with the sRBG color space usually. I've added the Alpha Component and
Color Space keys to each color definition as it's exported by iTerm.

See https://github.com/mbadolato/iTerm2-Color-Schemes/issues/106

This fixes all the colors in iTerm. E.g the red color in the dark
template was being rendered as #f6544f in sRGB, and now it's being
correctly rendered as #fa5750.

Also, I've updated the Selection Color to use bg_2 instead of bg_1, and
the Cursor Color to use fg_1 instead of fg_0, same as in other
templates, for better contrast.