Fix i3 configuration invalid and colors not read
Under i3status v2.13 the selenized config files are invalid, because the
hashtag `#` is interpreted as a start of a comment and not as the begin of a
color code.

This commit quotes the values to make sure i3status interpretes
these as colorcodes.
Currently i3 interprets the colorcodes correctly, but to make sure its
future proof the colorcodes in the i3 configs have been also quoted.

Example:

	`color_good = #70b433`
Is interpreted as comment, i3status throws an error.

	`color_good = "#70b433"`
i3status correctly interprets the colorcode.